+++ crates/acts/src/store/data/mod.rs
@@ -1,6 +1,7 @@
 mod event;
 mod message;
 mod model;
+mod op;
 mod package;
 mod proc;
 mod task;
@@ -8,6 +9,7 @@ mod task;
 pub use event::Event;
 pub use message::{Message, MessageStatus};
 pub use model::Model;
+pub use op::{Op, OpStatus, OpType};
 pub use package::Package;
 pub use proc::Proc;
 pub use task::Task;