+++ crates/acts/src/store/mod.rs
@@ -68,6 +68,8 @@ pub enum StoreIden {
     Messages,
     #[strum(serialize = "events")]
     Events,
+    #[strum(serialize = "ops")]
+    Ops,
 }
 
 #[derive(Debug, Deserialize, Serialize)]